pyproject.toml
[project]
name = "articulating-learned-rules"
version = "0.1.0"
description = "LLM rule articulation research project"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "matplotlib>=3.9.4",
    "seaborn>=0.13.2",
    "rouge-score>=0.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
